requirements.txt
celery[redis]
python-qpid-proton
pyzmq
anymarkup
